add logs fix ec2 instance termination

+++ utils/src/main/java/EC2Utils.java
@@ -1,11 +1,18 @@
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.ec2.Ec2Client;
 import software.amazon.awssdk.services.ec2.model.*;
+import software.amazon.awssdk.utils.CollectionUtils;
 
+import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class EC2Utils {
+    private final static Logger log = LoggerFactory.getLogger(EC2Utils.class);
+
     private final static Ec2Client ec2 = Ec2Client.builder()
             .region(Region.US_EAST_1)
             .build();
@@ -26,13 +33,14 @@ public class EC2Utils {
         RunInstancesRequest runRequest = RunInstancesRequest.builder()
                 .imageId(amiId)
                 .iamInstanceProfile(IamInstanceProfileSpecification.builder()
-                        .arn("arn:aws:iam::110380217222:instance-profile/assignment1")
+                        .arn("arn:aws:iam::882762034269:instance-profile/role1")
                         .build())
                 .instanceType(InstanceType.T2_MICRO)
                 .maxCount(instancesCount)
                 .minCount(instancesCount)
                 .userData(Base64.getEncoder().encodeToString(userDataScript.getBytes()))
-                .keyName("key1")
+                .keyName("dsp_key")
+                .securityGroups("default")
                 .build();
         RunInstancesResponse response = ec2.runInstances(runRequest);
         int numOfRetries = 0;
@@ -48,7 +56,7 @@ public class EC2Utils {
                     .build();
             try {
                 ec2.createTags(tagRequest);
-                System.out.println(
+                log.info(
                         "Successfully started EC2 instance: " + instanceId + "based on AMI: " + amiId);
 
             } catch (Ec2Exception e) {
@@ -56,7 +64,7 @@ public class EC2Utils {
                 i--;
                 numOfRetries++;
                 if (numOfRetries == 100) {
-                    System.out.println("EC2Utils.createEc2Instance() " +
+                    log.warn("EC2Utils.createEc2Instance() " +
                             "stop retrying tags creation for ec2 instances after 100 attempts");
                     break;
                 }
@@ -95,16 +103,48 @@ public class EC2Utils {
      * Terminate all running ec2 instances
      */
     public static void terminateEc2Instances() {
-        //todo: ask bar - if manager call this- doesnt it kills himself also? as it should.!
         String nextToken = null;
         do {
             DescribeInstancesRequest dRequest = DescribeInstancesRequest.builder().nextToken(nextToken).build();
             DescribeInstancesResponse response = ec2.describeInstances(dRequest);
+            List<String> workerInstanceIds = new ArrayList<>();
             for (Reservation reservation : response.reservations()) {
-                for (Instance instance : reservation.instances()) {
-                    TerminateInstancesRequest request = TerminateInstancesRequest.builder().instanceIds(instance.instanceId()).build();
-                    ec2.terminateInstances(request);
-                }
+                workerInstanceIds.addAll(reservation.instances().stream()
+                        .filter(ins -> !ins.tags().isEmpty()
+                                && ins.tags().stream().anyMatch(tag -> tag.key().toLowerCase().equals("name")
+                                && tag.value().toLowerCase().contains("worker"))
+                                && !ins.state().name().equals(InstanceStateName.TERMINATED)
+                        )
+                        .map(Instance::instanceId)
+                        .collect(Collectors.toList()));
+            }
+            if (!CollectionUtils.isNullOrEmpty(workerInstanceIds)) {
+                log.info("killing {} workers!", workerInstanceIds.size());
+                TerminateInstancesRequest request = TerminateInstancesRequest.builder()
+                        .instanceIds(workerInstanceIds)
+                        .build();
+                ec2.terminateInstances(request);
+                log.info("killed {} workers!", workerInstanceIds.size());
+            }
+            List<String> managerIds = new ArrayList<>();
+            for (Reservation reservation : response.reservations()) {
+                managerIds.addAll(reservation.instances().stream()
+                        .filter(ins -> !ins.tags().isEmpty()
+                                && ins.tags().stream().anyMatch(tag -> tag.key().toLowerCase().equals("name")
+                                && tag.value().toLowerCase().contains("manager"))
+                                && !ins.state().name().equals(InstanceStateName.TERMINATED)
+                        )
+                        .map(Instance::instanceId)
+                        .collect(Collectors.toList()));
+
+            }
+            if (!CollectionUtils.isNullOrEmpty(managerIds)) {
+                log.info("killing {} managers!", managerIds.size());
+                TerminateInstancesRequest request = TerminateInstancesRequest.builder()
+                        .instanceIds(managerIds)
+                        .build();
+                ec2.terminateInstances(request);
+                log.info("killed {} managers!", managerIds.size());
             }
             nextToken = response.nextToken();
         } while (nextToken != null);
